HAL: Fixed OOB read by assigning each param of Nancapabilities individually
NOTE: This is a fix for pa/2366563.

In pa/2366563, get_vendor_data_len() represents the size of
nan_hal_resp_t in the DHD not size of NanCapabilities.
It causes a crash issue again since memcpy's third argument
still represents 88 bytes which is defined in Legacy HAL.
(x2 register value - 0000000000000058)

Fixed it to assign each param individually instead of memcpy
regardless of framework NanCapabilities change.

Bug: 265225404
Test: BRCM halutil (brcm_halutil -nan -get_capabilities)

Change-Id: I7632d7f38f3c0a8ab2269b043eb77a17fb5100c8
Signed-off-by: Dennis Jeon <dennis.jeon@broadcom.corp-partner.google.com>